factor out reusable layout and charts for all population sections
