+++ src/population-viz/index.js
@@ -1,0 +1,1 @@
+export { default } from "./PopulationViz";
